+++ packages/kit/src/runtime/app/paths/client.js
@@ -1,6 +1,6 @@
 /** @import { Asset, RouteId, Pathname, ResolvedPathname } from '$app/types' */
 /** @import { ResolveArgs } from './types.js' */
-import { base, assets } from './internal/client.js';
+import { base, assets, hash_routing } from './internal/client.js';
 import { resolve_route } from '../../../utils/routing.js';
 
 /**
@@ -24,6 +24,8 @@ import { resolve_route } from '../../../utils/routing.js';
 export function asset(file) {
 	return (assets || base) + file;
 }
+
+const pathname_prefix = hash_routing ? '#' : '';
 
 /**
  * Resolve a pathname by prefixing it with the base path, if any, or resolve a route ID by populating dynamic segments with parameters.
@@ -51,7 +53,9 @@ export function asset(file) {
 export function resolve(...args) {
 	// The type error is correct here, and if someone doesn't pass params when they should there's a runtime error,
 	// but we don't want to adjust the internal resolve_route function to accept `undefined`, hence the type cast.
-	return base + resolve_route(args[0], /** @type {Record<string, string>} */ (args[1]));
+	return (
+		base + pathname_prefix + resolve_route(args[0], /** @type {Record<string, string>} */ (args[1]))
+	);
 }
 
 export { base, assets, resolve as resolveRoute };