+++ packages/kit/test/apps/basics/test/server.test.js
@@ -1364,13 +1364,6 @@ test.describe('tracing', () => {
 	});
 });
 
-test.describe('remote functions', () => {
-	test("doesn't write bundle to disk when treeshaking prerendered remote functions", () => {
-		test.skip(!!process.env.DEV, 'skip when in dev mode');
-		expect(fs.existsSync(path.join(root, 'dist'))).toBe(false);
-	});
-});
-
 test.describe('asset preload', () => {
 	if (!process.env.DEV) {
 		test('injects Link headers', async ({ request }) => {
@@ -1407,4 +1400,19 @@ test.describe('asset preload', () => {
 			expect(body).not.toContain('as="font"');
 		});
 	}
+});
+
+test.describe('Streaming', () => {
+	test("Discarded promises from server load functions don't hang SSR request", async ({
+		request
+	}) => {
+		let error;
+		try {
+			await request.get('/streaming/discarded-promise');
+		} catch (e) {
+			error = e;
+		}
+
+		expect(error).toBeUndefined();
+	});
 });